+++ test/fakePromise.js
@@ -1,0 +1,12 @@
+// --------------------
+// promisify-any module
+// FakePromise constructor used in tests
+// --------------------
+
+var FakePromise = module.exports = function() {};
+
+FakePromise.method = function() {
+    return function() {
+        return new FakePromise();
+    };
+};