+++ changelog.md
@@ -62,3 +62,4 @@
 
 * `.use()` method
 * Remove `.Promise` and `.co` properties
+* Remove `.defaultNumArgs` property