+++ changelog.md
@@ -64,3 +64,6 @@
 * Remove `.Promise` and `.co` properties
 * Remove `.defaultNumArgs` property
 * README
+
+Breaking changes:
+* To set the Promise implementation used, instead of setting `promisify.Promise` property, use `.use(Promise)`